backend-main/requirements.txt
asgiref==3.8.1
Django==4.2.18
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
Pillow==11.1.0
pip==24.3.1
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.1